change folder structure for deployment
