+++ .gitignore
@@ -1,0 +1,8 @@
+.env
+node_modules
+package-lock.json
+.idea
+lib/playground.js
+
+
+../.idea/
