--- backend/.gitignore
@@ -1,8 +1,0 @@
-.env
-node_modules
-package-lock.json
-.idea
-./lib/playground.js
-
-
-../.idea/
