--- backend/lib/cloudinary.config.js
@@ -1,10 +1,0 @@
-
-const cloudinary = require("cloudinary").v2
-
-cloudinary.config({
-    cloud_name: process.env.CLOUD_NAME,
-    api_key: process.env.API_KEY,
-    api_secret: process.env.API_SECRET
-})
-
-module.exports = cloudinary
