--- backend/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "scripts": {
-    "test": "jest",
-    "start": "node server",
-    "dev": "nodemon -r dotenv/config server"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcrypt": "^5.0.1",
-    "body-parser": "^1.19.0",
-    "cloudinary": "^1.25.1",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "jest": "^26.6.3",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.12.1",
-    "multer": "^1.4.2"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.7"
-  }
-}
