+++ lib/projectMockData.js
@@ -1,0 +1,11 @@
+const projectMockData = [{
+    title: 'Yahoo',
+    description: 'just a yahoo app..',
+    categories: ['news', 'search engine'],
+    bugs: ['ObjectId1', 'ObjectId2'], //bugSchema
+    members: ['ObjectId1', 'ObjectId2'], //userSchema
+    projectLead: 'ObjectId1' //userSchema
+}
+]
+
+module.exports = projectMockData
