+++ package.json
@@ -1,29 +1,32 @@
 {
-  "name": "bugs-are-us",
+  "name": "backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
+  "directories": {
+    "lib": "lib"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "cd backend && npm install & npm start"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/suha-nathan/bugs-are-us.git"
-  },
-  "engines": {
-    "node": "12.13.0"
+    "test": "jest",
+    "start": "node server",
+    "dev": "nodemon -r dotenv/config server"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/suha-nathan/bugs-are-us/issues"
+  "dependencies": {
+    "bcrypt": "^5.0.1",
+    "body-parser": "^1.19.0",
+    "cloudinary": "^1.25.1",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "jest": "^26.6.3",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.12.1",
+    "multer": "^1.4.2"
   },
-  "homepage": "https://github.com/suha-nathan/bugs-are-us#readme",
-  "engines": {
-    "node": "15.6.0",
-    "yarn": "1.22.5"
+  "devDependencies": {
+    "nodemon": "^2.0.7"
   }
-
 }
